requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
ansicolors==1.1.8         # via -r requirements.in
blis==0.7.4               # via spacy, thinc
catalogue==1.0.0          # via spacy, thinc
certifi==2020.12.5        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via nltk
colorama==0.4.4           # via crayons
configparser==5.0.1       # via webdriver-manager
crayons==0.4.0            # via webdriver-manager
cymem==2.0.5              # via preshed, spacy, thinc
datetime==4.3             # via -r requirements.in
idna==2.10                # via requests
importlib-metadata==3.1.1  # via catalogue
joblib==0.17.0            # via nltk, scikit-learn
murmurhash==1.0.5         # via preshed, spacy, thinc
nltk==3.5                 # via -r requirements.in
numpy==1.19.4             # via -r requirements.in, blis, pandas, scikit-learn, scipy, spacy, thinc
pandas==1.1.5             # via -r requirements.in
plac==1.1.3               # via spacy, thinc
preshed==3.0.5            # via spacy, thinc
prettytable==2.0.0        # via -r requirements.in
python-dateutil==2.8.1    # via pandas
python-dotenv==0.15.0     # via -r requirements.in
pytube==10.0.0            # via -r requirements.in
pytz==2020.4              # via datetime, pandas
regex==2020.11.13         # via nltk
requests==2.25.0          # via -r requirements.in, spacy, webdriver-manager
scikit-learn==0.23.2      # via sklearn
scipy==1.5.4              # via scikit-learn
selenium==3.141.0         # via -r requirements.in
six==1.15.0               # via python-dateutil
sklearn==0.0              # via -r requirements.in
spacy==2.3.5              # via -r requirements.in
srsly==1.0.5              # via spacy, thinc
thinc==7.4.5              # via spacy
threadpoolctl==2.1.0      # via scikit-learn
tqdm==4.54.1              # via -r requirements.in, nltk, spacy, thinc
typing-extensions==3.7.4.3  # via pytube
urllib3==1.26.2           # via requests, selenium
wasabi==0.8.0             # via spacy, thinc
wcwidth==0.2.5            # via prettytable
webdriver-manager==3.2.2  # via -r requirements.in
zipp==3.4.0               # via importlib-metadata
zope.interface==5.2.0     # via datetime

# The following packages are considered to be unsafe in a requirements file:
# setuptools
